+++ CHANGELOG.md
@@ -5,8 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
-------------
+[0.13.0] - 2022-12-23
+---------------------
 ### Added
 - Add [TypeScript](https://www.typescriptlang.org/) types.
 
@@ -87,7 +87,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/jbenner-radham/node-readme-md/compare/v0.12.0...HEAD
+[Unreleased]: https://github.com/jbenner-radham/node-readme-md/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.9.0...v0.10.0